pyproject.toml
[project]
name = "hr-analytics"
version = "1.0.0"
description = "People Analytics com IA — predição de attrition, insights com LLM e agente conversacional"
requires-python = ">=3.10"
dependencies = [
    # Dados
    "pandas>=2.0",
    "numpy>=1.24",
    "pyarrow>=14.0",
    "duckdb>=0.10",

    # ML
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "lightgbm>=4.0",
    "imbalanced-learn>=0.12",
    "shap>=0.45",

    # Otimização e tracking
    "optuna>=3.5",
    "mlflow[mcp]>=3.5.1",

    # API
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",

    # LLM
    "google-genai>=1.0",

    # Agente
    "langchain>=0.3",
    "langchain-google-genai>=2.0",

    # Frontend
    "streamlit>=1.35",
    "plotly>=5.18",
    "reportlab>=4.0",

    # Utilitários
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "joblib>=1.3",
    "tenacity>=8.2",
    "kagglehub>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "httpx>=0.27",
    "ruff>=0.3",
    "mypy>=1.8",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"src/hr_analytics/models/*" = ["N803", "N806", "E501"]  # X/y são convenção ML
"src/hr_analytics/inference/*" = ["N803", "N806"]
"src/hr_analytics/monitoring/*" = ["N803", "N806"]
"src/hr_analytics/agent/*" = ["E501"]  # prompts longos do LangChain
"tests/*" = ["N803", "N806", "E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
